refactor/cleanup/unit test/auto-lint/pre commit hooks
