+++ src/util/BluemixTerminal.ts
