+++ test/extension.test.ts
@@ -9,8 +9,8 @@ import * as assert from 'assert';
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
 import * as vscode from 'vscode';
-import {SystemCommand} from '../src/SystemCommand';
 import * as myExtension from '../src/extension';
+import {SystemCommand} from '../src/util/SystemCommand';
 
 // Defines a Mocha test suite to group tests of similar kind together
 suite('Extension Tests', () => {
@@ -32,7 +32,7 @@ suite('Extension Tests', () => {
 
 
     // this will fail on windows b/c its a linux/osx command
-    // testing a non-existant command execution 
+    // testing a non-existant command execution
     test('Error System Command', (done) => {
 
         let resultHandler = function(code) {
